+++ azext_edge/edge/providers/rpsaas/adr/asset_endpoint_profiles.py
@@ -4,7 +4,8 @@
 # Licensed under the MIT License. See License file in the project root for license information.
 # ----------------------------------------------------------------------------------------------
 
-from typing import Dict, List, Optional
+import json
+from typing import TYPE_CHECKING, Dict, Iterable, Optional, Union
 
 from knack.log import get_logger
 
@@ -12,135 +13,186 @@ from azure.cli.core.azclierror import (
     InvalidArgumentValueError,
     MutuallyExclusiveArgumentError,
     RequiredArgumentMissingError,
-    UnrecognizedArgumentError
 )
-from .base import ADRBaseProvider
 from .user_strings import (
     AUTH_REF_MISMATCH_ERROR,
-    CERT_AUTH_NOT_SUPPORTED,
     GENERAL_AUTH_REF_MISMATCH_ERROR,
     MISSING_USERPASS_REF_ERROR,
-    MISSING_TRANS_AUTH_PROP_ERROR,
     REMOVED_CERT_REF_MSG,
     REMOVED_USERPASS_REF_MSG,
-    UNRECOGNIZED_TRANS_AUTH_PROP_ERROR
 )
-from ....util import assemble_nargs_to_dict, build_query
-from ....common import ResourceTypeMapping, AEPAuthModes
+from ....util.az_client import get_registry_mgmt_client
+from ....util.queryable import Queryable
+from ....common import AEPAuthModes, AEPTypes
+
+if TYPE_CHECKING:
+    from ....vendor.clients.deviceregistrymgmt.operations import (
+        DiscoveredAssetEndpointProfilesOperations as DAEPOperations,
+        AssetEndpointProfilesOperations as AEPOperations
+    )
 
 logger = get_logger(__name__)
+AEP_RESOURCE_TYPE = "Microsoft.DeviceRegistry/assetEndpointProfiles"
+DISCOVERED_AEP_RESOURCE_TYPE = "Microsoft.DeviceRegistry/discoveredAssetEndpointProfiless"
 
 
-class AssetEndpointProfileProvider(ADRBaseProvider):
+# TODO: soul searching to see if I should combine with assets class
+class AssetEndpointProfiles(Queryable):
     def __init__(self, cmd):
-        super(AssetEndpointProfileProvider, self).__init__(
-            cmd=cmd,
-            resource_type=ResourceTypeMapping.asset_endpoint_profile.value,
+        super().__init__(cmd=cmd)
+        self.deviceregistry_mgmt_client = get_registry_mgmt_client(
+            subscription_id=self.default_subscription_id
         )
+        self.ops: "AEPOperations" = self.deviceregistry_mgmt_client.asset_endpoint_profiles
+        self.discovered_ops: "DAEPOperations" = self.deviceregistry_mgmt_client.discovered_asset_endpoint_profiles
+        self.update_ops: Optional[Union["AEPOperations", "DAEPOperations"]] = None
 
     def create(
         self,
         asset_endpoint_profile_name: str,
+        endpoint_profile_type: str,
+        instance_name: str,
         resource_group_name: str,
         target_address: str,
-        additional_configuration: Optional[str] = None,
-        certificate_reference: Optional[List[str]] = None,
-        cluster_name: Optional[str] = None,
-        cluster_resource_group: Optional[str] = None,
-        cluster_subscription: Optional[str] = None,
-        custom_location_name: Optional[str] = None,
-        custom_location_resource_group: Optional[str] = None,
-        custom_location_subscription: Optional[str] = None,
-        transport_authentication: Optional[str] = None,
+        certificate_reference: Optional[str] = None,
+        instance_resource_group: Optional[str] = None,
+        instance_subscription: Optional[str] = None,
         location: Optional[str] = None,
         password_reference: Optional[str] = None,
         username_reference: Optional[str] = None,
         tags: Optional[Dict[str, str]] = None,
+        discovered: bool = False,  # for easy discovered debugging
+        **additional_configuration
     ):
-        if certificate_reference:
-            raise InvalidArgumentValueError(CERT_AUTH_NOT_SUPPORTED)
-        extended_location = self.check_cluster_and_custom_location(
-            custom_location_name=custom_location_name,
-            custom_location_resource_group=custom_location_resource_group,
-            custom_location_subscription=custom_location_subscription,
-            cluster_name=cluster_name,
-            cluster_resource_group=cluster_resource_group,
-            cluster_subscription=cluster_subscription
+        from .helpers import get_extended_location
+        extended_location = get_extended_location(
+            cmd=self.cmd,
+            instance_name=instance_name,
+            instance_resource_group=instance_resource_group or resource_group_name,
+            instance_subscription=instance_subscription
         )
-
-        # Location
-        if not location:
-            location = self.get_location(resource_group_name)
+        cluster_location = extended_location.pop("cluster_location")
 
         auth_mode = None
         if not any([username_reference, password_reference, certificate_reference]):
             auth_mode = AEPAuthModes.anonymous.value
 
-        # Properties - bandaid for UI so it processes no transport auth correctly
-        properties = {"transportAuthentication": {"ownCertificates": []}}
+        # Properties
+        properties = {"endpointProfileType": endpoint_profile_type}
+
+        if endpoint_profile_type == AEPTypes.opcua.value:
+            properties["additionalConfiguration"] = _build_opcua_config(**additional_configuration)
+        # TODO: add other connector types in
         _update_properties(
             properties,
             target_address=target_address,
-            additional_configuration=additional_configuration,
             auth_mode=auth_mode,
             username_reference=username_reference,
             password_reference=password_reference,
             certificate_reference=certificate_reference,
-            transport_authentication=transport_authentication
         )
+        # discovered
+        if discovered:
+            self.ops = self.discovered_ops
+            properties.pop("authentication", None)
+            properties["version"] = 1
+            properties["discoveryId"] = "discoveryid1"
 
         aep_body = {
             "extendedLocation": extended_location,
-            "location": location,
+            "location": location or cluster_location,
             "properties": properties,
             "tags": tags,
         }
-        poller = self.resource_client.resources.begin_create_or_update(
-            resource_group_name=resource_group_name,
-            resource_provider_namespace=self.provider_namespace,
-            parent_resource_path="",
-            resource_type=self.resource_type,
-            resource_name=asset_endpoint_profile_name,
-            api_version=self.api_version,
-            parameters=aep_body
+        return self.ops.begin_create_or_replace(
+            resource_group_name,
+            asset_endpoint_profile_name,
+            resource=aep_body
         )
-        return poller
 
-    def query(
+    def delete(self, asset_endpoint_profile_name: str, resource_group_name: str):
+        self.show(
+            asset_endpoint_profile_name=asset_endpoint_profile_name,
+            resource_group_name=resource_group_name,
+            check_cluster=True
+        )
+        return self.update_ops.begin_delete(
+            resource_group_name,
+            asset_endpoint_profile_name,
+        )
+
+    def show(
+        self, asset_endpoint_profile_name: str, resource_group_name: str, check_cluster: bool = False
+    ) -> dict:
+        asset_endpoint = self.ops.get(
+            resource_group_name=resource_group_name, asset_endpoint_profile_name=asset_endpoint_profile_name
+        )
+        self.update_ops = self.ops
+        if check_cluster:
+            from .helpers import check_cluster_connectivity
+            check_cluster_connectivity(self.cmd, asset_endpoint)
+        return asset_endpoint
+
+    def list(self, resource_group_name: Optional[str] = None, discovered: bool = False) -> Iterable[dict]:
+        if discovered:
+            if resource_group_name:
+                return self.discovered_ops.list_by_resource_group(resource_group_name=resource_group_name)
+            return self.discovered_ops.list_by_subscription()
+
+        if resource_group_name:
+            return self.ops.list_by_resource_group(resource_group_name=resource_group_name)
+        return self.ops.list_by_subscription()
+
+    # TODO: unit test
+    def query_asset_endpoint_profiles(
         self,
-        additional_configuration: Optional[str] = None,
+        asset_endpoint_profile_name: Optional[str] = None,
         auth_mode: Optional[str] = None,
-        custom_location_name: Optional[str] = None,
+        custom_query: Optional[str] = None,
+        discovered: Optional[bool] = None,
+        endpoint_profile_type: Optional[str] = None,
+        instance_name: Optional[str] = None,
+        instance_resource_group: Optional[str] = None,
         location: Optional[str] = None,
         resource_group_name: Optional[str] = None,
         target_address: Optional[str] = None,
     ) -> dict:
-        query = ""
-        if additional_configuration:
-            query += f"| where properties.additionalConfiguration =~ \"{additional_configuration}\""
-        if auth_mode:
-            query += f"| where properties.userAuthentication.mode =~ \"{auth_mode}\""
-        if custom_location_name:
-            query += f"| where extendedLocation.name contains \"{custom_location_name}\""
-        if target_address:
-            query += f"| where properties.targetAddress =~ \"{target_address}\""
-
-        return build_query(
-            self.cmd,
-            subscription_id=self.subscription,
-            custom_query=query,
+        query_body = custom_query or _build_query_body(
+            asset_endpoint_profile_name=asset_endpoint_profile_name,
+            auth_mode=auth_mode,
+            endpoint_profile_type=endpoint_profile_type,
             location=location,
-            resource_group=resource_group_name,
-            type=ResourceTypeMapping.asset_endpoint_profile.full_resource_path,
-            additional_project="extendedLocation"
+            resource_group_name=resource_group_name,
+            target_address=target_address
         )
+
+        if discovered is not None:
+            resource_type = DISCOVERED_AEP_RESOURCE_TYPE if discovered else AEP_RESOURCE_TYPE
+            query = f"Resources | where type =~\"{resource_type}\" " + query_body
+        else:
+            # we put the query body into the each type query and then union to avoid the union result from
+            # becoming too big
+            query = f"Resources | where type =~ \"{AEP_RESOURCE_TYPE}\" {query_body} "\
+                f"| union (Resources | where type =~ \"{DISCOVERED_AEP_RESOURCE_TYPE}\" {query_body})"
+
+        if any([instance_name, instance_resource_group]):
+            instance_query = "Resources | where type =~ 'microsoft.iotoperations/instances' "
+            if instance_name:
+                instance_query += f"| where name =~ \"{instance_name}\""
+            if instance_resource_group:
+                instance_query += f"| where resourceGroup =~ \"{instance_resource_group}\""
+
+            # fetch the custom location + join on innerunique. Then remove the extra customLocation1 generated
+            query = f"{instance_query} | extend customLocation = tostring(extendedLocation.name) "\
+                f"| project customLocation | join kind=innerunique ({query}) on customLocation "\
+                "| project-away customLocation1"
+        return self.query(query=query)
 
     def update(
         self,
         asset_endpoint_profile_name: str,
         resource_group_name: str,
         target_address: Optional[str] = None,
-        additional_configuration: Optional[str] = None,
         auth_mode: Optional[str] = None,
         username_reference: Optional[str] = None,
         password_reference: Optional[str] = None,
@@ -151,7 +203,7 @@ class AssetEndpointProfileProvider(ADRBaseProvider):
         original_aep = self.show(
             asset_endpoint_profile_name,
             resource_group_name=resource_group_name,
-            check_cluster_connectivity=True
+            check_cluster=True
         )
         if tags:
             original_aep["tags"] = tags
@@ -161,102 +213,144 @@ class AssetEndpointProfileProvider(ADRBaseProvider):
         _update_properties(
             properties,
             target_address=target_address,
-            additional_configuration=additional_configuration,
             auth_mode=auth_mode,
             username_reference=username_reference,
             password_reference=password_reference,
             certificate_reference=certificate_reference,
         )
-
-        poller = self.resource_client.resources.begin_create_or_update_by_id(
-            resource_id=original_aep["id"],
-            api_version=self.api_version,
-            parameters=original_aep
-        )
-        return poller
-
-    def add_transport_auth(
-        self,
-        asset_endpoint_profile_name: str,
-        resource_group_name: str,
-        password_reference: str,
-        secret_reference: str,
-        thumbprint: str,
-    ):
-        original_aep = self.show(
+        # use this over update since we want to make sure we get the tags in
+        return self.update_ops.begin_create_or_replace(
+            resource_group_name,
             asset_endpoint_profile_name,
-            resource_group_name=resource_group_name,
-            check_cluster_connectivity=True
+            original_aep
         )
-        if original_aep["properties"].get("transportAuthentication") is None:
-            original_aep["properties"]["transportAuthentication"] = {
-                "ownCertificates": []
-            }
-
-        cert = {
-            "certThumbprint": thumbprint,
-            "certSecretReference": secret_reference,
-            "certPasswordReference": password_reference,
-        }
-        original_aep["properties"]["transportAuthentication"]["ownCertificates"].append(cert)
-
-        poller = self.resource_client.resources.begin_create_or_update_by_id(
-            resource_id=original_aep["id"],
-            api_version=self.api_version,
-            parameters=original_aep,
-        )
-        poller.wait()
-        original_aep = poller.result()
-        if not isinstance(original_aep, dict):
-            original_aep = original_aep.as_dict()
-        return original_aep["properties"]["transportAuthentication"]
-
-    def list_transport_auths(
-        self,
-        asset_endpoint_profile_name: str,
-        resource_group_name: str
-    ):
-        original_aep = self.show(
-            asset_endpoint_profile_name,
-            resource_group_name=resource_group_name
-        )
-
-        return original_aep["properties"].get("transportAuthentication", {"ownCertificates": []})
-
-    def remove_transport_auth(
-        self,
-        asset_endpoint_profile_name: str,
-        thumbprint: str,
-        resource_group_name: str,
-    ):
-        original_aep = self.show(
-            asset_endpoint_profile_name,
-            resource_group_name=resource_group_name,
-            check_cluster_connectivity=True
-        )
-        if original_aep["properties"].get("transportAuthentication") is None:
-            original_aep["properties"]["transportAuthentication"] = {
-                "ownCertificates": []
-            }
-
-        certs = original_aep["properties"]["transportAuthentication"]["ownCertificates"]
-        certs = [cert for cert in certs if cert["certThumbprint"] != thumbprint]
-
-        original_aep["properties"]["transportAuthentication"]["ownCertificates"] = certs
-
-        poller = self.resource_client.resources.begin_create_or_update_by_id(
-            resource_id=original_aep["id"],
-            api_version=self.api_version,
-            parameters=original_aep
-        )
-        poller.wait()
-        original_aep = poller.result()
-        if not isinstance(original_aep, dict):
-            original_aep = original_aep.as_dict()
-        return original_aep["properties"]["transportAuthentication"]
 
 
 # Helpers
+def _assert_above_min(param: str, value: int, minimum: int = 0) -> str:
+    if value < minimum:
+        return f"The parameter {param} needs to be at least {minimum}.\n"
+    return ""
+
+
+def _build_opcua_config(
+    original_config: Optional[str] = None,
+    application_name: Optional[str] = None,
+    auto_accept_untrusted_server_certs: Optional[bool] = None,
+    default_publishing_interval: Optional[int] = None,
+    default_sampling_interval: Optional[int] = None,
+    default_queue_size: Optional[int] = None,
+    keep_alive: Optional[int] = None,
+    run_asset_discovery: Optional[str] = None,
+    session_timeout: Optional[int] = None,
+    session_keep_alive: Optional[int] = None,
+    session_reconnect_period: Optional[int] = None,
+    session_reconnect_exponential_back_off: Optional[int] = None,
+    security_policy: Optional[str] = None,
+    security_mode: Optional[str] = None,
+    sub_max_items: Optional[int] = None,
+    sub_life_time: Optional[int] = None,
+) -> str:
+    config = json.loads(original_config) if original_config else {}
+
+    error_msg = ""
+    if application_name:
+        config["applicationName"] = application_name
+    if keep_alive:
+        error_msg += _assert_above_min("--keep-alive", keep_alive)
+        config["keepAliveMilliseconds"] = keep_alive
+    if run_asset_discovery is not None:
+        config["runAssetDiscovery"] = run_asset_discovery
+
+    # defaults
+    if any([
+        default_publishing_interval, default_sampling_interval, default_queue_size
+    ]) and not config.get("defaults"):
+        config["defaults"] = {}
+    if default_publishing_interval:
+        error_msg += _assert_above_min("--default-publishing-int", default_publishing_interval, -1)
+        config["defaults"]["publishingIntervalMilliseconds"] = default_publishing_interval
+    if default_sampling_interval:
+        error_msg += _assert_above_min("--default-sampling-int", default_sampling_interval, -1)
+        config["defaults"]["samplingIntervalMilliseconds"] = default_sampling_interval
+    if default_queue_size:
+        error_msg += _assert_above_min("--default-queue-size", default_queue_size, 0)
+        config["defaults"]["queueSize"] = default_queue_size
+
+    # session
+    if any([
+        session_timeout, session_reconnect_period, session_keep_alive, session_reconnect_exponential_back_off
+    ]) and not config.get("session"):
+        config["session"] = {}
+    if session_timeout:
+        error_msg += _assert_above_min("--session-timeout", session_timeout)
+        config["session"]["timeoutMilliseconds"] = session_timeout
+    if session_keep_alive:
+        error_msg += _assert_above_min("--session-keep-alive", session_keep_alive)
+        config["session"]["keepAliveIntervalMilliseconds"] = session_keep_alive
+    if session_reconnect_period:
+        error_msg += _assert_above_min("--session-reconnect-period", session_reconnect_period)
+        config["session"]["reconnectPeriodMilliseconds"] = session_reconnect_period
+    if session_reconnect_exponential_back_off:
+        error_msg += _assert_above_min("--session-reconnect-backoff", session_reconnect_exponential_back_off, -1)
+        config["session"]["reconnectExponentialBackOffMilliseconds"] = session_reconnect_exponential_back_off
+
+    # subscription
+    if any([sub_life_time, sub_max_items]) and not config.get("subscription"):
+        config["subscription"] = {}
+    if sub_life_time:
+        error_msg += _assert_above_min("--subscription-life-time", sub_life_time)
+        config["subscription"]["maxItems"] = sub_life_time
+    if sub_max_items:
+        error_msg += _assert_above_min("--subscription-max-items", sub_max_items, 1)
+        config["subscription"]["lifeTimeMilliseconds"] = sub_max_items
+
+    # security
+    if any([
+        auto_accept_untrusted_server_certs is not None, security_mode, security_policy
+    ]) and not config.get("security"):
+        config["security"] = {}
+    if auto_accept_untrusted_server_certs is not None:
+        config["security"]["autoAcceptUntrustedServerCertificates"] = auto_accept_untrusted_server_certs
+    if security_mode:
+        config["security"]["securityMode"] = security_mode
+    if security_policy:
+        config["security"]["securityPolicy"] = security_policy
+
+    if error_msg:
+        raise InvalidArgumentValueError(f"The following OPCUA connector arguments are invalid:\n {error_msg}")
+    return json.dumps(config)
+
+
+def _build_query_body(
+    asset_endpoint_profile_name: Optional[str] = None,
+    auth_mode: Optional[str] = None,
+    endpoint_profile_type: Optional[str] = None,
+    location: Optional[str] = None,
+    resource_group_name: Optional[str] = None,
+    target_address: Optional[str] = None,
+) -> str:
+    query_body = ""
+    if resource_group_name:
+        query_body += f"| where resourceGroup =~ \"{resource_group_name}\""
+    if location:
+        query_body += f"| where location =~ \"{location}\""
+    if asset_endpoint_profile_name:
+        query_body += f"| where name =~ \"{asset_endpoint_profile_name}\""
+    if auth_mode:
+        query_body += f"| where properties.authentication.method =~ \"{auth_mode}\""
+    if endpoint_profile_type:
+        query_body += f"| where properties.endpointProfileType =~ \"{endpoint_profile_type}\""
+    if target_address:
+        query_body += f"| where properties.targetAddress =~ \"{target_address}\""
+
+    query_body += "| extend customLocation = tostring(extendedLocation.name) "\
+        "| extend provisioningState = properties.provisioningState "\
+        "| project id, customLocation, location, name, resourceGroup, provisioningState, tags, "\
+        "type, subscriptionId "
+    return query_body
+
+
 def _process_authentication(
     auth_mode: Optional[str] = None,
     auth_props: Optional[Dict[str, str]] = None,
@@ -271,16 +365,16 @@ def _process_authentication(
         raise MutuallyExclusiveArgumentError(AUTH_REF_MISMATCH_ERROR)
 
     if certificate_reference and auth_mode in [None, AEPAuthModes.certificate.value]:
-        auth_props["mode"] = AEPAuthModes.certificate.value
-        auth_props["x509Credentials"] = {"certificateReference": certificate_reference}
+        auth_props["method"] = AEPAuthModes.certificate.value
+        auth_props["x509Credentials"] = {"certificateSecretName": certificate_reference}
         if auth_props.pop("usernamePasswordCredentials", None):
             logger.warning(REMOVED_USERPASS_REF_MSG)
     elif (username_reference or password_reference) and auth_mode in [None, AEPAuthModes.userpass.value]:
-        auth_props["mode"] = AEPAuthModes.userpass.value
+        auth_props["method"] = AEPAuthModes.userpass.value
         user_creds = auth_props.get("usernamePasswordCredentials", {})
-        user_creds["usernameReference"] = username_reference
-        user_creds["passwordReference"] = password_reference
-        if not all([user_creds["usernameReference"], user_creds["passwordReference"]]):
+        user_creds["usernameSecretName"] = username_reference
+        user_creds["passwordSecretName"] = password_reference
+        if not all([user_creds["usernameSecretName"], user_creds["passwordSecretName"]]):
             raise RequiredArgumentMissingError(MISSING_USERPASS_REF_ERROR)
         auth_props["usernamePasswordCredentials"] = user_creds
         if auth_props.pop("x509Credentials", None):
@@ -288,7 +382,7 @@ def _process_authentication(
     elif auth_mode == AEPAuthModes.anonymous.value and not any(
         [certificate_reference, username_reference, password_reference]
     ):
-        auth_props["mode"] = AEPAuthModes.anonymous.value
+        auth_props["method"] = AEPAuthModes.anonymous.value
         if auth_props.pop("x509Credentials", None):
             logger.warning(REMOVED_CERT_REF_MSG)
         if auth_props.pop("usernamePasswordCredentials", None):
@@ -299,28 +393,6 @@ def _process_authentication(
     return auth_props
 
 
-def _process_certificates(cert_list: Optional[List[List[str]]] = None) -> List[Dict[str, str]]:
-    """This is for the main create/update endpoint commands"""
-    if not cert_list:
-        return []
-    processed_certs = []
-    for cert in cert_list:
-        parsed_cert = assemble_nargs_to_dict(cert)
-        if not all(["thumbprint" in parsed_cert, "secret" in parsed_cert]):
-            raise RequiredArgumentMissingError(MISSING_TRANS_AUTH_PROP_ERROR.format(cert))
-        if not set(parsed_cert.keys()).issubset(set(["password", "thumbprint", "secret"])):
-            raise UnrecognizedArgumentError(UNRECOGNIZED_TRANS_AUTH_PROP_ERROR.format(cert))
-
-        processed_point = {
-            "certThumbprint": parsed_cert["thumbprint"],
-            "certSecretReference": parsed_cert["secret"],
-            "certPasswordReference": parsed_cert.get("password"),
-        }
-        processed_certs.append(processed_point)
-
-    return processed_certs
-
-
 def _update_properties(
     properties,
     target_address: Optional[str] = None,
@@ -329,19 +401,14 @@ def _update_properties(
     username_reference: Optional[str] = None,
     password_reference: Optional[str] = None,
     certificate_reference: Optional[str] = None,
-    transport_authentication: Optional[List[str]] = None,
 ):
     if additional_configuration:
         properties["additionalConfiguration"] = additional_configuration
     if target_address:
         properties["targetAddress"] = target_address
-    if transport_authentication:
-        properties["transportAuthentication"] = {
-            "ownCertificates": _process_certificates(transport_authentication)
-        }
     if any([auth_mode, username_reference, password_reference, certificate_reference]):
-        auth_props = properties.get("userAuthentication", {})
-        properties["userAuthentication"] = _process_authentication(
+        auth_props = properties.get("authentication", {})
+        properties["authentication"] = _process_authentication(
             auth_props=auth_props,
             auth_mode=auth_mode,
             certificate_reference=certificate_reference,
