+++ azext_edge/tests/edge/rpsaas/adr/conftest.py
@@ -66,7 +66,7 @@ def mocked_get_extended_location(mocker):
         "cluster_location": generate_random_string()
     }
     mock = mocker.patch(
-        "azext_edge.edge.providers.rpsaas.adr2.helpers.get_extended_location",
+        "azext_edge.edge.providers.rpsaas.adr.helpers.get_extended_location",
         return_value=result,
         autospec=True
     )
@@ -77,7 +77,7 @@ def mocked_get_extended_location(mocker):
 @pytest.fixture()
 def mocked_check_cluster_connectivity(mocker):
     yield mocker.patch(
-        "azext_edge.edge.providers.rpsaas.adr2.helpers.check_cluster_connectivity",
+        "azext_edge.edge.providers.rpsaas.adr.helpers.check_cluster_connectivity",
         autospec=True
     )
 
@@ -167,7 +167,7 @@ def get_profile_record(
 
 
 # Paths for mocking
-ASSETS_PATH = "azext_edge.edge.providers.rpsaas.adr2.assets"
+ASSETS_PATH = "azext_edge.edge.providers.rpsaas.adr.assets"
 
 # Generic objects
 # Assets
