+++ azext_edge/tests/edge/rpsaas/adr/test_asset_endpoint_profiles_int.py
